requirements.txt
flickrapi==2.1.2
PyYAML==3.11
pelican==3.6.3
# backport of Python 3.2+ configparser for 2.6 and 2.7
configparser; python_version <= '2.7'